+++ snake.h
@@ -1,6 +1,6 @@
 #ifndef SNAKE_H_INCLUDED
 #define SNAKE_H_INCLUDED
-const int menu_image_snake[200] = {
+const long menu_image_snake[200] = {
 	0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 	0, 0, 0, 0, 0, 1, 0, 0, 0, 0,
 	0, 0, 0, 0, 1, 1, 0, 0, 0, 0,
@@ -23,7 +23,7 @@ const int menu_image_snake[200] = {
 	0, 0, 0, 0, 0, 0, 0, 0, 0, 0
 };
 
-int init_snake(int** area, int** info_blocs);
+int init_snake(long** area, long** info_blocs);
 
-int next_step_snake(int* const area, int* const info_blocs, const SDL_Keycode* const key_code);
+int next_step_snake(long* const area, long* const info_blocs, const SDL_Keycode* const key_code);
 #endif
